(v0.1.3_3a) Inventory Tweaks - first part
Inventory has been tweaked in subtle ways that do not affect many of the game's systems. This includes removing the 2D nodes from the inventory slot, setting an aspectratiocontainer as the slot root node and fixing inventory scaling (only hotbar has been fixed so far, but it will be used as a base for all others). Committing now as I'm about to make some more hefty changes.