(v0.1.3_3b) Inventory Tweaks 2
Currently not working very well... but I have technically implemented an instance of the new inventory component as a hotbar. It's really looking like item dragging is gonna have to be a miss for this one. I might even get rid of the backpack for this project, to be honest.